extract `legendifyArtboard` logic to separate ditectory
