+++ Contents/Sketch/cleanUpPageLegends.js
@@ -1,8 +1,8 @@
-const isSketchStringsEqual = require('./isSketchStringsEqual');
+const isSketchStringsEqual = require('./utils/isSketchStringsEqual');
 const {
   LEGEND_ARTBOARD_NAME,
   LEGEND_ITEM_INDEX_NAME,
-} = require('../constants');
+} = require('./constants');
 
 function getLayersToRemove({ layer, itemsToRemove = [] }) {
   if (isSketchStringsEqual(layer.name(), LEGEND_ITEM_INDEX_NAME)) {
