+++ Contents/Sketch/legendifyArtboard/createLegendIndex.js
