+++ Contents/Sketch/legendifyArtboard/createLegendItem.js
