+++ Contents/Sketch/legendifyArtboard/createLegendItemIndexGenerator.js
