+++ Contents/Sketch/legendifyArtboard/index.js
@@ -1,0 +1,1 @@
+module.exports = require('./legendifyArtboard');
